+++ frontend/lib/models/submission_model.dart
@@ -1,0 +1,6 @@
+const List<String> submissionList = <String>[
+  'AppDev-51',
+  'AppDev-52',
+  'WebDev-41',
+  'WebDev-42'
+];
