+++ frontend/lib/pages/std_class.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class STD{
+class STD {
   final String _id;
 
   final String _name;
@@ -10,9 +10,6 @@ class STD{
   final String _linkedin;
 
   final String _position;
-
-
-
 
   STD(this._id, this._name, this._profilepic, this._linkedin, this._position);
 
@@ -25,6 +22,4 @@ class STD{
   String get linkedin => _linkedin;
 
   String get profilepic => _profilepic;
-
-
 }
